+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Add `small` prop and variant to `Tag`.
 - Remove truncation on `ProductMenu` and moves tag above the title.
 - Replace bottom `Feedback` component on examples pages with an `Aside`.
+- Allow passing children to the `ExamplesIndex` component.
 - Add theming options to `OverviewHeader`: `theme` to accept CSS classes to customize the container, `lightText` to enable white text, `description` to add a text description. The `features` prop is now optional.
 
 ## 1.3.0